New object uploaded. [ce_trickle] - 8/16/2023
[ce_trickle] is a way overly coded binary operator. 
This patch can be a bit redundant but can serve a purpose when multiple comparisons need to be checked with minimal code.

Bitwise operators will return a value once until a new or different end result is returned. So if the end result is the same regardless of the input the value will not pass.
The third outlet is not needed for bitwise operators and will return a -1 float.

ce_trickle provides the same functionality as all of this code below. Note* the end values are not the result of the binary operator but the corresponding values based on the floats provided. With the exception of bitwise operators as described above.